requirements.txt
streamlit==1.26.0
pandas==2.1.0
catboost==1.2.1.1
scikit-learn==1.3.0
scipy==1.11.2
seaborn
matplotlib==3.8.0
PyYAML==6.0.1
matplotlib==3.8.0
shap==0.42.1 